requirements.txt
Flask==2.3.0
pandas==2.0.0
scikit-learn==1.2.2
numpy==1.24.0
Pillow==10.0.0
pytesseract==0.3.10
python-dotenv==1.0.0
twilio==8.5.0
requests==2.31.0
pytesseract>=0.3.10
google-generativeai>=0.3.0
